what the student say about us and the eduSphere advantage added
the div that were on the hero section that were colored blue and blac k have been removed. and replaced with picture instead. and almost all elements width are changed to the use of % rather than px for responsiveness.a trasition-scale animation was added for the how it works section.